Runtime exception fixed

+++ src/main/kotlin/org/jetbrains/exposed/sql/ColumnType.kt
@@ -93,7 +93,7 @@ class CharacterColumnType : ColumnType() {
     override fun valueFromDB(value: Any): Any {
         return when(value) {
             is Char -> value
-            is Number -> value.toChar()
+            is Number -> value.toInt().toChar()
             is String -> value.single()
             else -> error("Unexpected value of type Char: $value")
         }
